[Fix] gcp인증키파일 경로 수정, 테마 전부 가져오기 수정, 지역분류 전부 가져오기 필요한지 체크해봐야함

+++ src/main/java/org/example/be/board/controller/BoardController.java
@@ -65,7 +65,7 @@ public class BoardController {
 
     // 게시글 수정
     @PostMapping("/{id}")
-    public ResponseEntity<CommonResponse<String>> update(@PathVariable int id,@ModelAttribute BoardDTO boardDTO) throws IOException {
+    public ResponseEntity<CommonResponse<String>> update(@PathVariable Long id,@ModelAttribute BoardDTO boardDTO) throws IOException {
         boardDTO.setId(id);
         boardService.updateBoard(boardDTO);
         return ResponseEntity.status(HttpStatus.CREATED)
